feat(web): 主题预设迁移至 theme 目录并扩展为整包配置（含 ElConf）
- 将 utils/themePreset.js 迁移为 src/theme/，内置预设拆分为 preset/*.json
  （含 name/version/order），用 import.meta.glob 自动发现并按 order 排序
- 预设由「仅视觉」扩展为整包系统配置（PRESET_KEYS 全量字段），BASE_CONFIG
  收口为唯一出厂默认；applyPreset 整包补齐覆盖，切换预设不再残留旧设置
- GVA 经典蓝 补全为完整原版配置，切换即还原 gin-vue-admin 出厂观感
- 新增一级 ElConf 注入 el-config-provider：size 由 config.global_size 派生、
  ElConf 显式声明优先（resolveElConfig），随预设整包保存/导入导出
- App.vue 以 v-bind="elConfig" 注入；presets 面板导出文案同步更新
