+++ web/src/App.vue
@@ -3,7 +3,7 @@
     id="app"
     class="bg-gray-50 text-slate-700 !dark:text-slate-500 dark:bg-slate-800"
   >
-    <el-config-provider :locale="zhCn" :size="appStore.config.global_size">
+    <el-config-provider :locale="zhCn" v-bind="appStore.elConfig">
       <router-view />
       <Application />
     </el-config-provider>
