+++ web/src/theme/preset/azir-fresh-blue.json
@@ -1,0 +1,35 @@
+{
+  "order": 2,
+  "version": 1,
+  "name": "Azir-清新蓝",
+  "builtin": true,
+  "config": {
+    "weakness": false,
+    "grey": false,
+    "primaryColor": "#5D87FF",
+    "successColor": "#60C041",
+    "warningColor": "#F9901F",
+    "dangerColor": "#F56C6C",
+    "infoColor": "#909399",
+    "showTabs": true,
+    "darkMode": "auto",
+    "layout_side_width": 256,
+    "layout_side_collapsed_width": 80,
+    "layout_side_item_height": 48,
+    "show_watermark": false,
+    "side_mode": "vertical",
+    "transition_type": "slide",
+    "global_size": "default",
+    "menu_theme": "design",
+    "radius": 0.75,
+    "card_mode": "shadow",
+    "show_breadcrumb": true,
+    "show_refresh": true,
+    "show_search": true,
+    "show_collapse_btn": true,
+    "header_bg": "rgba(255, 255, 255, 0)",
+    "header_border": "rgba(255, 255, 255, 0)",
+    "tabs_bg": "rgba(255, 255, 255, 0)"
+  },
+  "ElConf": {}
+}
