+++ web/src/theme/preset/gva-classic.json
@@ -1,0 +1,35 @@
+{
+  "order": 1,
+  "version": 2,
+  "name": "GVA 经典蓝",
+  "builtin": true,
+  "config": {
+    "weakness": false,
+    "grey": false,
+    "primaryColor": "#3b82f6",
+    "successColor": "#67c23a",
+    "warningColor": "#e6a23c",
+    "dangerColor": "#f56c6c",
+    "infoColor": "#909399",
+    "showTabs": true,
+    "darkMode": "auto",
+    "layout_side_width": 256,
+    "layout_side_collapsed_width": 80,
+    "layout_side_item_height": 48,
+    "show_watermark": true,
+    "side_mode": "normal",
+    "transition_type": "slide",
+    "global_size": "default",
+    "menu_theme": "light",
+    "radius": 0.125,
+    "card_mode": "border",
+    "show_breadcrumb": true,
+    "show_refresh": true,
+    "show_search": true,
+    "show_collapse_btn": true,
+    "header_bg": "",
+    "header_border": "",
+    "tabs_bg": ""
+  },
+  "ElConf": {}
+}
